+++ projects/_03_independent_cqc/_02_clean/tests/jobs/test_clean_ind_cqc_filled_posts.py
@@ -18,7 +18,6 @@ from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_data import (
 from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_schemas import (
     CleanIndCQCData as Schemas,
 )
-
 from utils import utils
 from utils.column_names.ind_cqc_pipeline_columns import (
     PartitionKeys as Keys,
