+++ projects/_03_independent_cqc/_02_clean/tests/utils/ascwds_filled_posts_calculator/test_ascwds_filled_posts_calculator.py
@@ -2,8 +2,12 @@ import unittest
 import warnings
 
 import projects._03_independent_cqc._02_clean.utils.ascwds_filled_posts_calculator.ascwds_filled_posts_calculator as job
-from tests.test_file_data import CalculateAscwdsFilledPostsData as Data
-from tests.test_file_schemas import CalculateAscwdsFilledPostsSchemas as Schemas
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_data import (
+    CalculateAscwdsFilledPostsData as Data,
+)
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_schemas import (
+    CalculateAscwdsFilledPostsSchemas as Schemas,
+)
 from utils import utils
 from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
