+++ projects/_03_independent_cqc/_02_clean/tests/utils/ascwds_filled_posts_calculator/test_utils.py
@@ -11,14 +11,14 @@ from pyspark.sql.types import (
 )
 
 import projects._03_independent_cqc._02_clean.utils.ascwds_filled_posts_calculator.utils as job
-from utils import utils
-from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
-from tests.test_file_data import (
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_data import (
     CalculateAscwdsFilledPostsUtilsData as Data,
 )
-from tests.test_file_schemas import (
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_schemas import (
     CalculateAscwdsFilledPostsUtilsSchemas as Schemas,
 )
+from utils import utils
+from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
 
 
 class TestAscwdsFilledPostsCalculatorUtils(unittest.TestCase):
